Updated the README
I also changed the name of the old README to weekthree. I could not think
of a better name or what casing to use. Feel free to change, please
read through the current README and make changes as you feel
necessary. Or feel free to tell me the changes youd like to see made.